requirements.txt
weaviate-client
transformers
pyyaml
pandas
torch
jieba
uvicorn
fastapi
scikit-learn
matplotlib
path